Search Affiliation Content on SCOPUS

Searches SCOPUS to get information about documents on an affiliation

affil_df(affil_id = NULL, affil_name = NULL, api_key = NULL,
  verbose = TRUE, facets = NULL, sort = "document-count", ...)

affil_data(affil_id = NULL, affil_name = NULL, api_key = NULL,
  verbose = TRUE, facets = NULL, sort = "document-count", ...)

Arguments

affil_id
Affiliation ID number.

affil_name
affiliation name. Disregarded if affil_id is specified

api_key
Elsevier API key

verbose
Print diagnostic messages

facets
Facets sent in query. See https://dev.elsevier.com/api_docs.html

sort
sorting sent to query

...
Arguments to be passed to author_search

Value

List of entries from SCOPUS

Note

The affil_data command will return the list of all entries as well as the data.frame

See also