Search Author Content on SCOPUS

Searches SCOPUS to get information about documents on an author. Note, author_list returns a list of the entries from author_search, but allows you to put in a name.

author_df(
  au_id = NULL,
  last_name = NULL,
  first_name = NULL,
  api_key = NULL,
  verbose = TRUE,
  all_author_info = FALSE,
  http = "https://api.elsevier.com/content/search/scopus",
  view = "COMPLETE",
  count = 25,
  general = TRUE,
  scrub = FALSE,
  headers = NULL,
  ...
)

author_df_orig(..., general = FALSE)

author_list(
  au_id = NULL,
  last_name = NULL,
  first_name = NULL,
  api_key = NULL,
  verbose = TRUE,
  http = "https://api.elsevier.com/content/search/scopus",
  view = "COMPLETE",
  count = 25,
  headers = NULL,
  ...
)

author_data(
  ...,
  verbose = TRUE,
  all_author_info = FALSE,
  general = TRUE,
  scrub = FALSE
)

Arguments

au_id

Author ID number. Overrides any first/last name argument

last_name

last name of author

first_name

first name of author

api_key

Elsevier API key

verbose

Print diagnostic messages

all_author_info

Should all author info be recorded instead of that just to the author given

http

Address for scopus api

view

type of view to give, see https://api.elsevier.com/documentation/ScopusSearchAPI.wadl

count

number of records to retrieve (below 25, see https://dev.elsevier.com/api_key_settings.html)

general

Should gen_entries_to_df instead of the way before version 0.5.10.9001

scrub

Should `scrub_identifier` be run on the identifier?

headers

Headers passed to add_headers, passed to GET

...

Arguments to be passed to author_search

Value

List of entries from SCOPUS

Note

The author_data command will return the list of all entries as well as the data.frame.

See also

Examples

if (is_elsevier_authorized()) {
res = author_df(last_name = "Muschelli", first_name = "John",
verbose = FALSE)
}