Search Author Content on SCOPUS

Searches SCOPUS to get information about documents on an author.

author_search(
  au_id,
  api_key = NULL,
  http = "https://api.elsevier.com/content/search/author",
  count = 200,
  start = 0,
  verbose = TRUE,
  facets = "subjarea(sort=fd,count=350)",
  searcher = "AU-ID",
  max_count = Inf,
  view = c("STANDARD", "COMPLETE"),
  add_query = NULL,
  headers = NULL,
  wait_time = 0,
  ...
)

Arguments

au_id
Author ID number

api_key
API Key for Elsevier

http
Address for scopus api

count
number of records to retrieve (below 200, see https://dev.elsevier.com/api_key_settings.html)

start
where should the records start gathering

verbose
Print diagnostic messages

facets
Facets sent in query. See https://dev.elsevier.com/api_docs.html

searcher
Identifier for author ID. Do not change unless you know exactly what the API calls for.

max_count
Maximum count of records to be returned.

view
type of view to give, see https://api.elsevier.com/documentation/AuthorSearchAPI.wadl

add_query
Things to add to the query parameter for the request

headers
additional headers to be added to add_headers

wait_time
The time in seconds to wait across consecutive requests of a single search (when records > 25)

...
Arguments to be passed to the query list for GET

Value

List of entries from SCOPUS

See also

Examples

if (FALSE) {
author_search(au_id = "Smith", searcher = "affil(princeton) and authlast")
berk = author_search(au_id = "berkeley", searcher = "affil", count =100)
}