Search Author Content on SCOPUS

Searches SCOPUS to get information about documents on an author.

author_search(au_id, api_key = NULL,
  http = "http://api.elsevier.com/content/search/scopus", count = 25,
  verbose = TRUE, facets = "subjarea(sort=fd)", searcher = "AU-ID",
  max_count = Inf, ...)

Arguments

au_id
Author ID number

api_key
API Key for Elsevier

http
Address for scopus api

count
number of records to retrieve (below 25, see http://dev.elsevier.com/api_key_settings.html)

verbose
Print diagnostic messages

facets
Facets sent in query. See http://dev.elsevier.com/api_docs.html

searcher
Identifer for author ID. Do not change unless you know exactly what the API calls for.

max_count
Maximum count of records to be returned.

...
Arguments to be passed to the query list for GET

Value

List of entries from SCOPUS

See also

get_author_info